FIRST COMMIT
This commit is a first commit of what I had worked on over the last bit of September. This in particular is a simpler version of GUI-based to-do
list I attempted to make where the user can input command-line commands to create a to-do list, add to it, remove from it, and show it. I also
added help functionality to print all commands and their descriptions.

Where I got stuck on this project, and would like to return in the future, is file saving and loading so that lists can be loaded and saved.